temp_install/pyproject.toml
[project]
name = "socialseed-tasker"
version = "0.1.0"
description = "A graph-based task management framework for AI agents"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "SocialSeed"}
]
keywords = ["task-management", "neo4j", "graph-database", "ai-agents", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # CLI
    "typer>=0.9.0",
    "rich>=13.0.0",
    
    # API
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    
    # Validation
    "pydantic>=2.5.0",
    
    # Database
    "neo4j>=5.15.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "testcontainers[neo4j]>=3.7.0",
    
    # Linting
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    
    # Type stubs
    "types-setuptools",
    
    # Documentation
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["socialseed_tasker"]

[tool.setuptools.package-data]
socialseed_tasker = [
    "../.agent/**",
]

[project.scripts]
ss-tasker = "socialseed_tasker.entrypoints.terminal_cli.app:main_entry"

[tool.ruff]
target-version = "py310"
line-length = 120
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.ruff.isort]
known-first-party = ["socialseed_tasker"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[[tool.mypy.overrides]]
module = ["neo4j.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=socialseed_tasker --cov-report=term-missing"